zoom to element capability
Set option directZoomToElement under zoom to true for a map.
After this you`re able to directly zoom to an element via doubleclick
and to zoom out doubleclick the same element.
Works in combination with mousewheel zoom functionality as it saves the
current zoom level.
Animation has been used to zoom to an element (see animateviewbox
function which uses raphaels easing functions)

Real World Example: I had 25 locations in germany and 50 locations in
europe - to have an overview and to be able to hover to the correct
locations I needed this.

This works only for SVG (which has been implemented as condition to work
only for SVG) !
VML and raphael have an issue with getBBox() -> wrong values will be
taken ... should be fixed in the next release.

Result:
Works for
IE 9 / 10 / 11
Firefox
Chrome

Is deactivated for:
IE 6 - 8