Tooltip position computation as absolute
With tooltip set as absolute positioning, the Top and Left computation needs to be changed.
The tooltip (absolute div) is inside the map (relative div), hence the Top/Left is relative to the upper left corner of the map div.
This simplifies the computation : we need the pointer localisation (e.page...) and substract the container offset ($container.offset()...).
Of course, we take into accounts when the tooltip is near the container edge. Hence we take the minimum between the container edge minus the tooltip size and the mouse positioning.