JSHint: moved function out of loop
Fixes 2 JSHint warning: `Don't make functions within a loop.`
I found that the loop to delete the plots is quite similar to the one to delete links. Hence, I created a single function `fnRemovePlot()` which is used by each.
After that, I noticed the same pattern for the loop to show the new plots and the new links. Hence I created a new function `fnShowPlot()` which is used by both.
Code rationalization! Yiaie!